+++ src/lib/handler/settingshandler.h
@@ -197,8 +197,8 @@ public:
     static bool getLinkToRelatedAxisChecked(QString channel);
     static void setLinkToRelatedAxisChecked(QString channel, bool value);
     static void setLinkToRelatedAxis(QString channel, QString linkedChannel);
-    static void setDelayValue(QString channel, int value);
-    static int getDelayValue(QString channel);
+    static void setDelayValue(QString channel, float value);
+    static float getDelayValue(QString channel);
 
     static void setLibraryView(int value);
     static void setThumbSize(int value);
